requirements.txt
Flask>=3.0.3
PyYAML>=6.0.1
requests>=2.32.3
rcon>=2.1.1
tcping==0.1.1rc1
gevent>=24.2.1
prettytable>=0.2.13